"Fixed" error with ctypes not loading on Android
Kodi running on Android doesn't have the attribute:
ctypes.pythonapi.PyCell_New . Not sure exactly what functionality it
provides to dill, but everything seems to still be working without it.
Code now checks if running on android and doesn't load it if that's
indeed the case. Possibly a performance hit.