Fixed issue Collection TV showing again when browsing add-ons
Collection TV started showing again on root when browsing add-ons . This
happened due to new addon.xml setting.

The add-on is marked as also providing audio, images, and executable (as
opposed to just videos). Due to this Kodi sends extra parameters in the
URL so the add-on can identify where it was accessed from. This change
made the recognition check fail.